python/requirements.txt
langchain>=1.0.0a0
langchain-anthropic>=1.0.0a0
langchain-openai>=1.0.0a0
langsmith>=0.3.0
anthropic>=0.65.0
openai>=2.0.0
python-dotenv>=0.19.0
httpx>=0.24.0
pydantic>=2.0.0

